+++ NEWS.md
@@ -1,6 +1,8 @@
 # httr2 (development version)
 
-* `resp_body_json/html/xml()` now work for cached responses (@mgirlich, #277).
+* Cached responses now combine the headers of the new response with the headers
+  of the cached response. In particular, this fixes `resp_body_json/xml/html()`
+  on cached responses (@mgirlich, #277).
 
 * `with_mock()` and `local_mock()` now correctly trigger errors when the
   mocked response represents an HTTP failure (#252).